Moved code output into separate component
